move mocked stdout tests to `tests/stdio.rs`
